Remove the sub folder withFilter
